Move components to other directory
